+++ .formatter.exs
@@ -1,0 +1,6 @@
+[
+  locals_without_parens: [assert_value: :*],
+  export: [
+    locals_without_parens: [assert_value: :*]
+  ]
+]