+++ test/diff_test.exs
@@ -30,20 +30,55 @@ defmodule AssertValue.DiffTest do
      fff
     +GGG
     """
+
+    assert diff(nil, b) == """
+    -
+    +aaa
+    +xxx
+    +ddd
+    +eee
+    +fff
+    +GGG
+    """
+    assert diff(a, nil) == """
+    -aaa
+    -bbb
+    -ccc
+    -ddd
+    -eee
+    -fff
+    +
+    """
+  end
+
+  test "string and integer" do
+    assert diff(42, "42") == """
+    -42
+    +\"42\"
+    """
+    assert diff("42", 42) == """
+    -\"42\"
+    +42
+    """
+  end
+
+  test "nil and integer" do
+    assert diff(42, nil) == """
+    -42
+    +
+    """
+    assert diff(nil, 42) == """
+    -
+    +42
+    """
   end
 
   test "char list diff" do
     a = 'aaa\nbbb\nccc\nddd\neee\nfff\n'
     b = "aaa\nxxx\nddd\neee\nfff\nGGG\n"
     assert diff(a, b) == """
-     aaa
-    -bbb
-    -ccc
-    +xxx
-     ddd
-     eee
-     fff
-    +GGG
+    -'aaa\\nbbb\\nccc\\nddd\\neee\\nfff\\n'
+    +\"aaa\\nxxx\\nddd\\neee\\nfff\\nGGG\\n\"
     """
   end